pyproject.toml
# See the NOTICE file distributed with this work for additional information
# regarding copyright ownership.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
# http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

# pyproject.toml -- MEvoLib project configuration

## Generic configuration
[project]
name = "MEvoLib"
dynamic = [
    "version",
]
requires-python = ">= 3.8"
description = "Molecular Evolution Library for Python"
readme = "README.md"
authors = [
    {name = "Jorge Alvarez-Jarreta", email = "jalvarez@ebi.ac.uk"}
]
license = {text = "Apache License 2.0"}
keywords = [
    "bioinformatics",
    "phylogenetics",
    "molecular evolution",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "biopython >= 1.79",
    "dendropy >= 4.5.2",
    "numpy >= 1.23.2",
]

[project.optional-dependencies]
dev = [
    "mypy",
    "pylint",
    "pytest",
    "Sphinx",
]

[project.urls]
repository = "https://github.com/JAlvarezJarreta/mevolib"
paper = "https://bmcbioinformatics.biomedcentral.com/articles/10.1186/s12859-016-1303-3"

[project.scripts]
fetch_genbank_seqs = "mevolib.fetch.BioSeqs:main"
get_genes = "mevolib.cluster:main"
get_align = "mevolib.align:main"

[build-system]
requires = [
    "setuptools",
    "setuptools-scm",
    "wheel"
]
build-backend = "setuptools.build_meta"

[tool.setuptools.dynamic]
version = {attr = "mevolib.__version__"}

[tool.pylint.messages_control]
max-line-length = 110
disable = [
    "invalid-name",
    "similarities",
    "too-few-public-methods",
    "too-many-arguments",
    "too-many-locals",
    "too-many-statements",
    "unspecified-encoding",
    "wildcard-import"
]

[tool.mypy]
ignore_missing_imports = true
warn_unused_configs = true

[tool.black]
line-length = 110